+++ buildapp.js
@@ -1,10 +1,14 @@
 #!/usr/bin/env node
 const path = require('path');
 const {spawn} = require('child_process');
+// 运行命令的路径
 const runtimePath = process.cwd();
+// build.js所在的路径
 const codePath = __dirname;
+// ctbuild.cmd或者ctbuild.sh所在路径
 const commandPath = path.join(codePath, 'node_modules', '.bin', path.sep);
-let argsMap;
+// 配置文件所在路径
+let configPath;
 
 /**
  * corssenvTask
@@ -33,41 +37,41 @@ function corssenvTask() {
   });
 }
 
-/**
- * prodDllTask
- * @return {Promise}
- */
-function prodDllTask() {
-  return new Promise((resolve, reject) => {
-    const command = process.platform === "win32" ? `${commandPath}webpack.cmd` : `${commandPath}webpack`;
-    const devDllProcess = spawn(
-      command,
-      [
-        '--config',
-        path.join('webpackconfig', 'webpack.prod.dll.js'),//'webpackconfig/webpack.prod.dll.js',
-        '--custom',
-        path.join(runtimePath, path.sep),//`${runtimePath}\\`
-      ],
-      {
-        cwd: codePath,
-        encoding: 'utf-8',
-      }
-    );
-
-    devDllProcess.stdout.on('data', (data) => {
-      console.log(`stdout: ${data}`);
-    });
-
-    devDllProcess.stderr.on('data', (data) => {
-      console.log(`stderr: ${data}`);
-    });
-
-    devDllProcess.on('close', (code) => {
-      console.log(`prodDllTaskClose：${code}`);
-      resolve();
-    });
-  });
-}
+// /**
+//  * prodDllTask
+//  * @return {Promise}
+//  */
+// function prodDllTask() {
+//   return new Promise((resolve, reject) => {
+//     const command = process.platform === "win32" ? `${commandPath}webpack.cmd` : `${commandPath}webpack`;
+//     const devDllProcess = spawn(
+//       command,
+//       [
+//         '--config',
+//         path.join('webpackconfig', 'webpack.prod.dll.js'),//'webpackconfig/webpack.prod.dll.js',
+//         '--custom',
+//         path.join(runtimePath, path.sep),//`${runtimePath}\\`
+//       ],
+//       {
+//         cwd: codePath,
+//         encoding: 'utf-8',
+//       }
+//     );
+//
+//     devDllProcess.stdout.on('data', (data) => {
+//       console.log(`stdout: ${data}`);
+//     });
+//
+//     devDllProcess.stderr.on('data', (data) => {
+//       console.log(`stderr: ${data}`);
+//     });
+//
+//     devDllProcess.on('close', (code) => {
+//       console.log(`prodDllTaskClose：${code}`);
+//       resolve();
+//     });
+//   });
+// }
 
 /**
  * webpackTask
@@ -81,13 +85,13 @@ function webpackTask() {
       [
         '--open',
         '--config',
-        path.join('webpackconfig', 'webpack.prod.js'),//'webpackconfig/webpack.prod.js',
+        path.join('webpackconfig', 'webpack.prod.js'),
         '--progress',
         '--colors',
         '--runtimepath',
-        path.join(runtimePath, '/'),//`${runtimePath}\\`,
+        path.join(runtimePath, path.sep),
         '--customconfig',
-        argsMap.get('--config')
+        configPath
       ],
       {
         cwd: codePath,
@@ -139,8 +143,22 @@ function loopTask() {
 }
 
 module.exports = {
-  build: (args) => {
-    argsMap = args;
+  /**
+   * build
+   * @param {String} - ctbuildConfigPath
+   * ctbuild.config.js配置文件的路径，如果没有指定则会寻找命令运行目录下的ctbuild.config.js文件
+   */
+  build: (ctbuildConfigPath = '') => {
+    if(ctbuildConfigPath) {
+      if(path.isAbsolute(ctbuildConfigPath)) {
+        configPath = ctbuildConfigPath;
+      } else {
+        configPath = path.join(runtimePath, ctbuildConfigPath);
+      }
+    } else {
+      configPath = path.join(runtimePath,'ctbuild.config.js');
+    }
+
     loopTask().then(() => {
       console.log('finish');
       process.exit();