+++ gulpfile.js
@@ -5,22 +5,19 @@ const sourceMap = require('gulp-sourcemaps');
 const copyexts = ['less', 'svg', 'jpg', 'jpeg', 'png', 'bmp', 'json', 'eot', 'woff', 'ttf'];
 const commandArgs = require('./commandArgs');
 const argsMap = commandArgs.initCommandArgs();
-const runtimePath = argsMap.get('--runtimepath')[0];
-const srcPath = argsMap.get('--srcpath')[0];
+const outputpath = argsMap.get('--outputpath')[0];
+const compilePath = argsMap.get('--compilepath')[0];
 
 // buildpackage生成的目录名称
 const generateDirName = 'lib';
-// buildpackage原始名称
-const srcDirName = 'src';
 
 /**
  * copy
  */
 gulp.task('copy', () => {
   for (let i = 0; i < copyexts.length; i++) {
-    // gulp.src(`${srcPath}${srcDirName}\\**\\*.${copyexts[i]}`).pipe(gulp.dest(`${runtimePath}${generateDirName}`));
-    gulp.src(path.join(`${srcPath}${srcDirName}`,'**',`*.${copyexts[i]}`))
-        .pipe(gulp.dest(`${runtimePath}${generateDirName}`));
+    gulp.src(path.join(compilePath,'**',`*.${copyexts[i]}`))
+        .pipe(gulp.dest(outputpath));
   }
 });
 
@@ -31,12 +28,12 @@ gulp.task('minjs', () => {
   return gulp.src([
     // `${runtimePath}lib\\**\\*.js`,
     // `${runtimePath}lib\\**\\*.jsx`,
-    path.join(`${runtimePath}${generateDirName}`,'**','*.js'),
-    path.join(`${runtimePath}${generateDirName}`,'**','*.jsx'),
+    path.join(outputpath,'**','*.js'),
+    path.join(outputpath,'**','*.jsx'),
   ]).pipe(sourceMap.init())
     .pipe(uglify())
     .pipe(sourceMap.write('.'))
-    .pipe(gulp.dest(`${runtimePath}${generateDirName}`))
+    .pipe(gulp.dest(outputpath))
 });
 
 gulp.task('default', ['copy', 'minjs']);